Refactor all yes/no option imports to use common/constants
