+++ src/apps/transformers.js
@@ -2,7 +2,7 @@
 const { filter, upperFirst } = require('lodash')
 
 const { format, isDateValid } = require('../client/utils/date')
-const { OPTION_NO, OPTION_YES } = require('./constants')
+const { OPTION_NO, OPTION_YES } = require('../common/constants')
 const { hqLabels } = require('./companies/labels')
 const groupExportCountries = require('../lib/group-export-countries')
