+++ src/client/modules/Companies/AccountManagement/Objective/ObjectiveForm.jsx
@@ -3,7 +3,11 @@ import { Button } from 'govuk-react'
 import Link from '@govuk-react/link'
 
 import { TASK_SAVE_OBJECTIVE } from '../state'
-import { FORM_LAYOUT } from '../../../../../common/constants'
+import {
+  FORM_LAYOUT,
+  OPTIONS_YES_NO,
+  OPTION_YES,
+} from '../../../../../common/constants'
 import {
   DefaultLayout,
   Form,
@@ -16,7 +20,6 @@ import {
 } from '../../../../components'
 import urls from '../../../../../lib/urls'
 import { buildCompanyBreadcrumbs } from '../../utils'
-import { OPTIONS_YES_NO, OPTION_YES } from '../../../../../apps/constants'
 import { GREY_3, TEXT_COLOUR } from '../../../../utils/colours'
 
 const ButtonSecondary = (props) => (
